split the animated encoder into its own module
The frame decoders and the GIF encode share nothing with the still path
but the size to scale to and where to store the result, and no behavior
changes here.
